emberly-mvp/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
firebase-admin==6.8.0
pymongo==4.13.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.1.0
pydantic>=1.7.4,<3.0.0
starlette>=0.40.0,<0.47.0
pytest==7.3.1
httpx==0.24.0
boto3==1.26.129
websockets==11.0.2
pytest-asyncio==0.21.0
# PostgreSQL dependencies
sqlalchemy==2.0.41
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.16.1
greenlet==3.2.2
# Additional dependencies
motor==3.7.1
pydantic-settings==2.9.1
email-validator==2.2.0 